Implement Incomplete Cholesky preconditioner
This signficantly reduces the number of iterations required to converge
on an answer during the pressure solve. Aside from generally better
performance, this also dramatically reduces the likelyhood of hitting
the iteration limit. That's good because inaccurate solutions were the
primary cause of some weird artifacts, like the color spike beneath the
fluid source in the waterfall scenario.