Improve fluid rendering resolution
I originally rendered the fluid using one simulation grid cell per
character. I used [o,O,0] to represent varying amounts of marker
particles in a cell. However, terminal characters are typically twice
as tall as they are wide, so the simulation would get stretched
vertically during rendering.

One way of addressing this would be to change the simulation to use
tall rectanglular cells rather than square cells. In theory, this would
not be too difficult. I only tried this briefly, but the results did not
look as good as I hoped. I'm unsure if that wass due to a bug or if it
was the correct behaviour, but in either case it seemed like significant
effort to track down the cause.

Instead, I have changed the input and output to the simulation. I use
Unicode block characters and set the forground or background colours
to render two grid cells in each character. The input file format now
also interprets each character as representing two cells.

This means that the program can now display simulations with double
the previous vertical resolution! Though, on my laptop the frame rate
sometimes drops significantly with those large grids. This project has
tried to lean towards minimal abstraction and readability over
performance. That remains the case, though I've introduced a few
abstractions as I've tried to ensure correctness. I might now also need
to optimize a little to ensure it runs well on screen-sized simulations.