Improve marker updates
Ensure that fluid isn't flowing out-of-bounds or into solids, strip
a few unecessary comments, and use a few equivalent but shorter calls.

The check on g_solid should really be an assertion, but perfect
collision-handling is hard and my improved marker collision-handling
requires small steps that seem to exacerbate other problems.